Services gesplitst in interface en implementation class
