+++ src/main/java/nl/novi/catsittermanager/exceptions/BadRequestException.java
@@ -1,7 +1,5 @@
 package nl.novi.catsittermanager.exceptions;
 
-import nl.novi.catsittermanager.services.UserService;
-
 public class BadRequestException extends RuntimeException {
     private static final long serialVersionUID = 1L;
     public BadRequestException() {
