+++ src/main/java/nl/novi/catsittermanager/services/UserService.java
@@ -1,93 +1,4 @@
 package nl.novi.catsittermanager.services;
 
-import nl.novi.catsittermanager.dtos.user.UserDto;
-import nl.novi.catsittermanager.exceptions.UsernameNotFoundException;
-import nl.novi.catsittermanager.models.Authority;
-import nl.novi.catsittermanager.models.User;
-import nl.novi.catsittermanager.repositories.UserRepository;
-import nl.novi.catsittermanager.utils.RandomStringGenerator;
-import nl.novi.catsittermanager.mappers.UserMapper.fromUser;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-
-@Service
-public class UserService {
-    private final UserRepository userRepository;
-    private PasswordEncoder encoder;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    public List<UserDto> getUsers() {
-        List<UserDto> collection = new ArrayList<>();
-        List<User> list = userRepository.findAll();
-        for (User user : list) {
-            collection.add(fromUser(user));
-        }
-        return collection;
-    }
-
-    public UserDto getUser(String username) {
-        UserDto dto = new UserDto();
-        Optional<User> user = userRepository.findById(username);
-        if (user.isPresent()){
-            dto = fromUser(user.get());
-        }else {
-            throw new UsernameNotFoundException(username);
-        }
-        return dto;
-    }
-
-    public boolean userExists(String username) {
-        return userRepository.existsById(username);
-    }
-
-    public String createUser(UserDto userDto) {
-        String randomString = RandomStringGenerator.generateAlphaNumeric(20);
-        encoder.encode(userDto.password);
-        User newUser = userRepository.save(toUser(userDto));
-        return newUser.getUsername();
-    }
-
-    public void deleteUser(String username) {
-        userRepository.deleteById(username);
-    }
-
-    public void updateUser(String username, UserDto newUser) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
-        User user = userRepository.findById(username).get();
-        user.setPassword(newUser.password);
-        userRepository.save(user);
-    }
-
-    public Set<Authority> getAuthorities(String username) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
-        User user = userRepository.findById(username).get();
-        UserDto userDto = fromUser(user);
-        return userDto.authorities;
-    }
-
-    public void addAuthority(String username, String authority) {
-
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
-        User user = userRepository.findById(username).get();
-        user.addAuthority(new Authority(username, authority));
-        userRepository.save(user);
-    }
-
-    public void removeAuthority(String username, String authority) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
-        User user = userRepository.findById(username).get();
-        Authority authorityToRemove = user.getAuthorities().stream().filter((a) -> a.getAuthority().equalsIgnoreCase(authority)).findAny().get();
-        user.removeAuthority(authorityToRemove);
-        userRepository.save(user);
-    }
-
+public interface UserService {
 }
